+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ```
     * Do not call updateOrder on SDK pre-payment method.
     * Enable Nuvei GW for Zero-Total Orders, but allow only Credit Cards.
+    * Disable DCC when Order total is Zero.
 ```
 
 # 1.5